requirements.txt
future==0.18.2
iso8601==0.1.13
paho-mqtt==1.5.1
PyYAML==5.3.1
pyserial==3.4
redis==3.5.3